Last child of scrollable has margin. Chat padding.
The chat had its padding re-added so that it's the default that semantic
gives it.

Added a style that adds a small margin to the bottom of the last child
of a scrollable element. This makes it so that the content doesn't look
cut off, and the user knows they've reached the bottom.