Auto play videos. Catch AbortError from promise.
When player.destroy calls detachMediaElement, Firefox throws an
AbortError. From looking at the flv.js source, detachMediaElement sets
the source of the stream to ''. It's possible Firefox is attempting to
refetch the stream from the '' url, but playback has already been
aborted. This is just my guess.

For now, simply catch the AbortError and set player to null, hoping that
the garbage collector successfully cleans up.